FIXED Implicit cast warning

+++ src/exec/rule/command/init_redirection.c
@@ -31,7 +31,7 @@ int exec_do_redirect_double_left(const char *word)
     char *line = NULL;
     size_t len = 0;
     int fd[2];
-    int ret = 1;
+    ssize_t ret = 1;
 
     if (pipe(fd) == -1) {
         perror("pipe");
@@ -40,7 +40,7 @@ int exec_do_redirect_double_left(const char *word)
     write(1, "? ", 2);
     ret = getline(&line, &len, stdin);
     for (; ret > 0 && strcmp(line, word) != '\n'; ret = getline(&line, &len, stdin)) {
-        write(fd[1], line, ret);
+        write(fd[1], line, (size_t) ret);
         write(1, "? ", 2);
     }
     if (ret < 0)
